universal-shorts/requirements.txt
openai
requests
google-auth
google-auth-oauthlib
google-api-python-client
pillow
gtts
